+++ Code/CrowdControlMod.cs
@@ -955,6 +955,7 @@ public sealed class CrowdControlMod : Mod
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Food));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Kite));
         AddEffect(new SpawnPiggyBankEffect());
+        AddEffect(new SpawnVoidVaultEffect());
 
         // --- World effects
         AddEffect(new UseDialEffect(true));